# src/auth/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/groups": {
            "get": {
                "tags": [
                    "组"
                ],
                "summary": "列出所有组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "过滤条件",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "页数",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "页尺寸",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"groups\":[{\"id\":1,\"name\":\"group1\",\"description\":\"group1\",\"created_at\":\"2021-01-21 11:20:29\",\"updated_at\":\"2021-01-21 11:20:29\"}],\"message\":\"Success\",\"page\":1,\"page_size\":50,\"pages\":1,\"total\":1}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "组"
                ],
                "summary": "新建组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.Group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\",\"group\":{\"id\":1,\"name\":\"group1\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/groups/{group_id}": {
            "put": {
                "tags": [
                    "组"
                ],
                "summary": "更新组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "组ID",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.Group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\",\"group\":{\"id\":1,\"name\":\"group_rename\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "组"
                ],
                "summary": "删除组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "组ID",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/groups/{group_id}/users": {
            "get": {
                "tags": [
                    "组"
                ],
                "summary": "列出角色所有用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "组ID",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\",\"users\":[{\"id\":4,\"username\":\"test2\",\"is_owner\":false,\"joined_at\":\"2021-01-20 11:01:16\"},{\"id\":3,\"username\":\"test\",\"is_owner\":true,\"joined_at\":\"2021-01-20 11:01:32\"}]}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "组"
                ],
                "summary": "添加用户至组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "组ID",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.UserGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/groups/{group_id}/users/{user_id}": {
            "put": {
                "tags": [
                    "组"
                ],
                "summary": "设置用户是否是组Owner",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "组ID",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.IsOwnerForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "组"
                ],
                "summary": "移除组中用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "组ID",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "database.Group": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "database.UserGroup": {
            "type": "object",
            "required": [
                "is_owner",
                "user_id"
            ],
            "properties": {
                "is_owner": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "form.IsOwnerForm": {
            "type": "object",
            "required": [
                "is_owner"
            ],
            "properties": {
                "is_owner": {
                    "type": "boolean"
                }
            }
        },
...
    }
}